requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
pinecone-client
cohere
python-dotenv
tiktoken
pypdf